requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anthropic==0.28.1
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==42.0.8
daphne==4.1.2
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.14
diff-match-patch==20230430
dirtyjson==1.0.8
distro==1.9.0
dj-rest-auth==6.0.0
Django==4.2
django-cors-headers==4.3.1
django-filter==24.2
django-import-export==4.0.7
django-timezone-field==7.0
djangorestframework==3.15.1
drf-api-tracking==1.8.4
drf-yasg==1.21.7
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
jiter==0.5.0
joblib==1.4.2
kombu==5.3.7
marshmallow==3.21.3
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pandas==2.2.2
pillow==10.3.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
PyJWT==2.9.0
pyOpenSSL==24.1.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.3
redis==5.0.6
regex==2024.5.15
requests==2.31.0
requests-oauthlib==2.0.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
sqlparse==0.5.0
tablib==3.5.0
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tqdm==4.66.4
Twisted==24.3.0
txaio==23.1.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.9.4
zope.interface==6.4.post2
pysqlite3-binary==0.5.2.post3
django-allauth==0.60.0
